Reorganize of screens
