+++ src/screens/main_screens/main_tabs/Products.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { 
     Text,
 } from "react-native";
-import { getUserToken } from "../../AuthMethods";
+import { getUserToken } from "../../../AuthMethods";
 import axios from 'axios';
 import Error from './screens/Error'
 import Loading from './screens/Loading'
@@ -82,7 +82,11 @@ class Products extends Component {
                         />     
             }
             else{
-                return <AllProducts/>
+                return <AllProducts
+                            onPressVendorPhoto={() => {this.props.navigation.navigate('ProfileDetails')}}
+                            onPressCard={() => {this.props.navigation.navigate('ProductDetails')}}
+                
+                        />
             }
         }
     }
